requirements.txt
numpy==1.22.4
torch==1.13.1
Pillow==9.1.1
torchvision
matplotlib==3.5.2
tqdm==4.64.0
imageio==2.19.2
imageio-ffmpeg==0.4.7
click==8.1.3
easydict==1.9
PyYAML==6.0
kmeans-pytorch
wandb
datetime
scipy
numpy
GPUtil
opencv-python
git+https://github.com/openai/CLIP.git